# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: entries
    description: Journal entry operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /summarize-bodily-recording:
    post:
      operationId: summarizeBodilyRecording
      summary: Summarize bodily symptoms recording
      description: Summarizes user's spoken bodily symptoms for a specific body part using Groq AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bodyPart:
                  type: string
                transcript:
                  type: string
              required:
                - bodyPart
                - transcript
      responses:
        "200":
          description: Summary generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SummarizeResponse"
  /entries:
    get:
      operationId: getEntries
      tags: [entries]
      summary: Get all journal entries for a user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: All entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntriesResponse"
  /entries/{date}:
    put:
      operationId: saveEntry
      tags: [entries]
      summary: Save or update a journal entry
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalEntry"
      responses:
        "200":
          description: Entry saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
    delete:
      operationId: deleteEntry
      tags: [entries]
      summary: Delete a journal entry
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Entry deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteEntryResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    SummarizeResponse:
      type: object
      properties:
        summary:
          type: string
      required:
        - summary
    BodyMetrics:
      type: object
      properties:
        energy:
          type: number
        tension:
          type: number
        sleep:
          type: number
        hydration:
          type: number
      required:
        - energy
        - tension
        - sleep
        - hydration
    PainMarker:
      type: object
      properties:
        slug:
          type: string
        painLevel:
          type: number
        notes:
          type: string
        summary:
          type: string
      required:
        - slug
        - painLevel
        - notes
        - summary
    JournalEntry:
      type: object
      properties:
        id:
          type: string
        date:
          type: string
        prompt:
          type: string
        response:
          type: string
        mood:
          type: number
        bodyMetrics:
          $ref: "#/components/schemas/BodyMetrics"
        createdAt:
          type: string
        painMarkers:
          type: array
          items:
            $ref: "#/components/schemas/PainMarker"
      required:
        - id
        - date
        - prompt
        - response
        - mood
        - bodyMetrics
        - createdAt
    EntriesResponse:
      type: object
      properties:
        entries:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/JournalEntry"
      required:
        - entries
    DeleteEntryResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok