# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * Summarizes user's spoken bodily symptoms for a specific body part using Groq AI
 * @summary Summarize bodily symptoms recording
 */
export const SummarizeBodilyRecordingBody = zod.object({
  "bodyPart": zod.string(),
  "transcript": zod.string()
})

export const SummarizeBodilyRecordingResponse = zod.object({
  "summary": zod.string()
})


/**
 * @summary Get all journal entries for a user
 */
export const GetEntriesQueryParams = zod.object({
  "userId": zod.coerce.string()
})

export const GetEntriesResponse = zod.object({
  "entries": zod.record(zod.string(), zod.object({
  "id": zod.string(),
  "date": zod.string(),
  "prompt": zod.string(),
  "response": zod.string(),
  "mood": zod.number(),
  "bodyMetrics": zod.object({
  "energy": zod.number(),
  "tension": zod.number(),
  "sleep": zod.number(),
  "hydration": zod.number()
}),
  "createdAt": zod.string(),
  "painMarkers": zod.array(zod.object({
  "slug": zod.string(),
  "painLevel": zod.number(),
  "notes": zod.string(),
  "summary": zod.string()
})).optional()
}))
})


/**
 * @summary Save or update a journal entry
 */
export const SaveEntryParams = zod.object({
  "date": zod.coerce.string()
})

export const SaveEntryQueryParams = zod.object({
  "userId": zod.coerce.string()
})

export const SaveEntryBody = zod.object({
  "id": zod.string(),
  "date": zod.string(),
  "prompt": zod.string(),
  "response": zod.string(),
  "mood": zod.number(),
  "bodyMetrics": zod.object({
  "energy": zod.number(),
  "tension": zod.number(),
  "sleep": zod.number(),
  "hydration": zod.number()
}),
  "createdAt": zod.string(),
  "painMarkers": zod.array(zod.object({
  "slug": zod.string(),
  "painLevel": zod.number(),
  "notes": zod.string(),
  "summary": zod.string()
})).optional()
})

export const SaveEntryResponse = zod.object({
  "id": zod.string(),
  "date": zod.string(),
  "prompt": zod.string(),
  "response": zod.string(),
  "mood": zod.number(),
  "bodyMetrics": zod.object({
  "energy": zod.number(),
  "tension": zod.number(),
  "sleep": zod.number(),
  "hydration": zod.number()
}),
  "createdAt": zod.string(),
  "painMarkers": zod.array(zod.object({
  "slug": zod.string(),
  "painLevel": zod.number(),
  "notes": zod.string(),
  "summary": zod.string()
})).optional()
})


/**
 * @summary Delete a journal entry
 */
export const DeleteEntryParams = zod.object({
  "date": zod.coerce.string()
})

export const DeleteEntryQueryParams = zod.object({
  "userId": zod.coerce.string()
})

export const DeleteEntryResponse = zod.object({
  "ok": zod.boolean()
})